cleanup

+++ scripts/clean.mjs
@@ -3,9 +3,17 @@ import * as Fs from "node:fs";
 
 const dirs = [".", ...Glob.sync("packages/*/")];
 dirs.forEach((pkg) => {
-    const files = [".tsbuildinfo", "build", "dist", "coverage"];
+    const files = [".tsbuildinfo", "tsconfig.tsbuildinfo", "docs", "build", "dist", "coverage"];
 
     files.forEach((file) => {
+        if (pkg === "." && file === "docs") {
+            return;
+        }
+
         Fs.rmSync(`${pkg}/${file}`, { recursive: true, force: true }, () => {});
     });
+});
+
+Glob.sync("docs/*/").forEach((dir) => {
+    Fs.rmSync(dir, { recursive: true, force: true }, () => {});
 });
